refactor: PH2-T06 move WebSocket ticker into manja-ticker
Context:

- Phase: PH2
- Task: PH2-T06 Move WebSocket ticker into manja-ticker

Changes:

- Added ticker client, stream, and models modules to
  manja-ticker/src/{client.rs,stream.rs,models/**} and wired crate
  exports via manja-ticker/src/lib.rs
- Declared WebSocket runtime dependencies (tokio-tungstenite,
  tungstenite, stubborn-io, tracing, serde, serde_json,
  secrecy, futures-util, manja-core) in
  manja-ticker/Cargo.toml and integrated UserSession from
  manja_core::models in manja-ticker/src/stream.rs
- Updated manja/Cargo.toml to depend on manja-ticker as an
  optional crate and changed the websocket feature to enable
  dep:manja-ticker instead of direct WebSocket dependencies
- Replaced the in-crate ticker implementation under
  manja/src/kite/ticker/** with a thin facade in
  manja/src/kite/ticker/mod.rs that re-exports
  KiteStreamCredentials, StreamState, Mode, TickerRequest,
  TickerStream, and WebSocketClient from manja-ticker
- Adjusted manja/src/main.rs to log ticker StreamState using
  to_uri() and removed the direct tungstenite import now provided
  via manja-ticker
- Preserved the public ticker API used by
  manja/examples/ticker.rs by keeping the
  `manja::kite::ticker::{KiteStreamCredentials, Mode, StreamState,

Notes:

- Tests: cargo build --all (default features)
- Follow-up: add focused ticker tests in manja-ticker and update
  docs/examples per PH2-T08/PH2-T09
