+++ manja-ticker/Cargo.toml
@@ -7,4 +7,13 @@ description = "WebSocket ticker client for the manja SDK"
 repository = "https://github.com/saxena-dev/manja"
 
 [dependencies]
-
+futures-util = "0.3.30"
+manja-core = { path = "../manja-core" }
+secrecy = "0.8.0"
+serde = { version = "1.0.202", features = ["derive"] }
+serde_json = "1.0.117"
+stubborn-io = "0.3.5"
+tokio = { version = "1.37.0", features = ["full"] }
+tokio-tungstenite = { version = "0.23.1", features = ["native-tls"] }
+tracing = "0.1.40"
+tungstenite = "0.23.0"
