+++ manja/Cargo.toml
@@ -36,26 +36,22 @@ secrecy = "0.8.0"
 serde = { version = "1.0.202", features = ["derive"] }
 serde_json = "1.0.117"
 sha2 = "0.10.8"
-stubborn-io = { version = "0.3.5", optional = true }
 thiserror = "1.0.61"
 tokio = { version = "1.37.0", features = ["full"] }
-tokio-tungstenite = { version = "0.23.1", features = ["native-tls"], optional = true }
 totp-rs = { version = "5.5.1", optional = true }
 tracing = "0.1.40"
 tracing-subscriber = "0.3.18"
-tungstenite = { version = "0.23.0", optional = true }
 url = "2.5.0"
 manja-core = { path = "../manja-core" }
 manja-http = { path = "../manja-http" }
+manja-ticker = { path = "../manja-ticker", optional = true }
 
 [features]
 default = ["websocket", "backoff", "dotenv-config", "webdriver-login"]
 
 # Enable WebSocket-based ticker client and its dependencies.
 websocket = [
-    "dep:tokio-tungstenite",
-    "dep:tungstenite",
-    "dep:stubborn-io",
+    "dep:manja-ticker",
 ]
 
 # Enable WebDriver-based login flow (Chrome automation + TOTP).
